# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DataFlow API
  description: Secure, self-hosted family data vault API
  version: 0.2.0
  license:
    name: MIT
servers:
  - url: http://localhost:3460
    description: Development server
paths:
...
  /api/auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, master_password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                master_password:
                  type: string
                display_name:
                  type: string
      responses:
        '201':
          description: User registered
  /api/auth/login:
    post:
      summary: Log in
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, master_password]
              properties:
                email:
                  type: string
                password:
                  type: string
                master_password:
                  type: string
      responses:
        '200':
          description: Login successful
...
  /api/categories/{id}:
    get:
      summary: Get a category
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category details
    put:
      summary: Update a category
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category updated
    delete:
      summary: Delete a category
      tags: [Categories]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Category deleted
...
  /api/record-types/{id}:
    get:
      summary: Get a record type
      tags: [Record Types]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Record type details
    put:
      summary: Update a record type
      tags: [Record Types]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Record type updated
    delete:
      summary: Delete a record type
      tags: [Record Types]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Record type deleted
...
  /api/items:
    get:
      summary: List items
      tags: [Items]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: category_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of items
    post:
      summary: Create an item
      tags: [Items]
      responses:
        '201':
          description: Item created
...
  /api/items/{id}:
    get:
      summary: Get an item
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item details
    put:
      summary: Update an item
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item updated
    delete:
      summary: Delete an item (soft delete)
      tags: [Items]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item deleted
...
  /api/tags/{id}:
    put:
      summary: Update a tag
      tags: [Tags]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tag updated
    delete:
      summary: Delete a tag
      tags: [Tags]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tag deleted
...
  /api/settings/{key}:
    put:
      summary: Set a setting
      tags: [Settings]
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Setting updated
    delete:
      summary: Delete a setting
      tags: [Settings]
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Setting deleted
...
  /api/members/{id}:
    get:
      summary: Get member details
      tags: [Members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Member details
    put:
      summary: Update member
      tags: [Members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Member updated
    delete:
      summary: Delete member (admin only)
      tags: [Members]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Member deleted
...
  /api/emergency/{id}:
    delete:
      summary: Cancel emergency access request
      tags: [Emergency]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Request cancelled
...